backend/requirements.txt
altair==5.5.0
attrs==25.4.0
beautifulsoup4==4.12.2
blinker==1.9.0
bs4==0.0.2
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
et_xmlfile==2.0.0
gitdb==4.0.12
GitPython==3.1.41
GooglePatentsPdfDownloader @ git+https://github.com/lorenzbr/GooglePatentsPdfDownloader.git@944f830b7e9033d389aeffac53105b37a79cab1b
h11==0.16.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lxml==6.0.2
MarkupSafe==3.0.3
narwhals==2.10.2
numpy==2.3.4
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pluggy==1.6.0
protobuf==6.33.0
pyarrow==21.0.0
pycparser==2.23
pydeck==0.9.1
Pygments==2.19.2
pypdf==4.3.1
PyPDF2==3.0.1
PySocks==1.7.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
reportlab==4.0.9
requests==2.32.5
rpds-py==0.28.0
selenium==4.15.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
streamlit==1.51.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
webdriver-manager==4.0.0
websocket-client==1.9.0
wsproto==1.3.2
